+++ rnd/autogpt_server/Dockerfile.ws
@@ -25,10 +25,10 @@ ENV POETRY_VERSION=1.8.3 \
     PATH="$POETRY_HOME/bin:$PATH"
 RUN pip3 install poetry
 
-COPY rnd/autogpt_server /app/rnd/autogpt_server
-COPY rnd/autogpt_libs /app/rnd/autogpt_libs
 COPY autogpt /app/autogpt
 COPY forge /app/forge
+COPY rnd/autogpt_libs /app/rnd/autogpt_libs
+COPY rnd/autogpt_server /app/rnd/autogpt_server
 
 WORKDIR /app/rnd/autogpt_server
