+++ rnd/autogpt_server/pyproject.toml
@@ -44,13 +44,6 @@ uvicorn = { extras = ["standard"], version = "^0.30.1" }
 websockets = "^12.0"
 youtube-transcript-api = "^0.6.2"
 
-ollama = "^0.3.0"
-feedparser = "^6.0.11"
-python-dotenv = "^1.0.1"
-expiringdict = "^1.2.2"
-discord-py = "^2.4.0"
-
-autogpt-libs = {path = "../autogpt_libs"}
 aio-pika = "^9.4.3"
 redis = "^5.0.8"
